ad_video_generator/requirements.txt
streamlit==1.54.0
uvicorn==0.34.0
fastapi==0.115.8
pydantic==2.10.6
requests==2.32.5
moviepy==2.2.1
pillow==12.1.1
numpy==2.4.2
edge-tts==7.0.0
python-multipart==0.0.20